Change scramble() to generate safer characters
 ## Problem
`scramble()` currently replaces non-alphanumeric ascii characters with different ascii character. The problem with this, is that the replacement character could cause the string to no longer be valid. For example:
* If we scramble an email, we often end up including urls that are not valid in an email address
* If we scramble a url that has encoded chars, we often end up making that url unsafe

 ## Solution
As a simple solution for now: when replacing non-alphanumeric ascii characters, use a range of characters that is generally safer (e.g. for emails and urls): alpha-numeric characters, or one of `-`, `_`, `+`.

We could always have more specific scramble functions in the future (e.g. `copycat.scramble.email()` or `copycat.scramble.url()`), but at least for now, I think it would be helpful to use a more conservative replacement character range for the general case.

The one caveat I can think of, is that the output range is narrower than the input range, so the outputs will be less unique. As far as collisions go, I've been meaning to investigate how likely collisions are more generally for copycat, planning on finding out more about this then.