chore: Add basic collision probabilities script
 ## Context
We currently have very little information on how likely collisions are for each of copycat's API methods. For some methods (e.g. bool), these are trivial or unimportant. For data types that typically need to be unique (e.g. `uuid` or `email`), this is important information to know: it allows us to know where in the API we need to improve to make collisions less likely, and lets our users know how likely collisions will be for their use case.

 ## Worth noting
Methods like `scramble` are difficult to measure this for, since the probability is largely dependant on the input string. For example, an input string of a short length will have a higher collision rate. For this reason, `scramble` is omitted from the script at the moment.

 ## Approach
* For each API method:
  * While we're under the minimum number of runs (100), or until the margin of error is below 0.05% if we're still under the maximum number of runs (9999):
    * Run the method with a new uuid v4 as input until the first collision is found
  * Output the obtained stats (e.g. mean, stddev, margin of error)

Since the task here is number-crunching, we do this using a worker farm to leverage multiple cpus to get results faster.

 ## How to read the results
For each api method the `mean` is roughly: the data set size at which the first collision is likely to happen.

`runs` is the number of times we tested the api method before deciding the margin of error is "good enough" - under 5%.

"likely to happen" is pretty vague though. What the numbers are really saying, is: we are 95% certain that the "real" first collision data set size is within 5% of the `mean` displayed in these results. Unless we reached the maximum number of runs - 9999, in which case, it is whatever the `moe` stat shows instead of `0.05` (5%).

 ## Limitations
I must admit that I know very little about this subject (hash collision probabilities). That, and I didn't pay enough attention in my stats lectures back in the day. I don't know where these measurements suck, or how badly they do. Any input would be greatly appreciated!

 ## Alternatives and why I didn't go for them
An alternative approach would be to test for the probability of collisions at a particular data set size (similar to how you might do this for calculating numbers for the birthday problem). This is arguably less helpful for our case, since for each api method, we want to know _when_ the first collision happened, so that we know how useful the copycat api methods would be for some given data set size (e.g. if email's first collision is likely to happen for a data set size of 100,000 and we need uniqueness, then we have some surety that it is fine to use for a data set of size 500, but not one of size 900,000).

 ## Results
I haven't run this for all API methods yet, it takes a while to run. I don't know if its broken for some methods yet, and will need to run it for longer to get results for the other API methods.

```
{"methodName":"bool","mean":"1.00","stddev":"0.00","moe":"0.00","runs":2}
{"methodName":"digit","mean":"6.00","stddev":"0.00","moe":"0.00","runs":2}
{"methodName":"oneOf","mean":"2.00","stddev":"0.00","moe":"0.00","runs":2}
{"methodName":"char","mean":"9.49","stddev":"4.84","moe":"0.05","runs":401}
{"methodName":"hex","mean":"4.90","stddev":"2.33","moe":"0.05","runs":349}
{"methodName":"lastName","mean":"26.77","stddev":"14.68","moe":"0.05","runs":463}
{"methodName":"timezone","mean":"12.56","stddev":"6.61","moe":"0.05","runs":427}
{"methodName":"word","mean":"19.32","stddev":"9.81","moe":"0.05","runs":396}
{"methodName":"dateString","mean":"421.42","stddev":"216.10","moe":"0.05","runs":406}
{"methodName":"email","mean":"98728.50","stddev":"3043.50","moe":"0.04","runs":2}
{"methodName":"country","mean":"19.50","stddev":"9.61","moe":"0.05","runs":374}
```

An interesting one is `email` - only 2 runs is a bit dodgy - I'll need to look into that one further.