feat: Support scrambling other types (#35)
Currently, only scrambling of strings are supported as inputs to scramble. This PR adds support for scrambling values of other types:

If a number is given, each digit will be replaced, and the floating point (if relevant) will be preserved:

```js
copycat.scramble(782364.902374)
// => 387998.531441
```

If an object or array is given, the values inside the object or array will be recursively scrambled:

```js
copycat.scramble({
  a: [
    {
      b: 23,
      c: 'foo',
    },
  ],
})
/* =>
{
  "a": [{
    "b": 19,
    "c": "ebp"
  }]
}
*/
```

If a date is given, each segment in the date will be scrambled:

```js
copycat.scramble(new Date(89723948723948)))
// => Date(7754-02-02T15:03:19.862Z)
```

If a boolean or null value is given, the value will simply be returned.