fix!: Improve float()'s distribution (via fictional upgrade 0.10.1 -> 1.0.0) (#55)
From https://github.com/oftherivier/fictional/pull/36:

When a `min` and `max` is given, `float()` tends to stick around the `min`.

This PR changes the implementation to ensure the values distribute more evenly across the (min, max) range of values.

**BREAKING CHANGE**
Inputs for `float()` will now map to a different output after upgrading fictional. It is still deterministic, the same input will always map to the same output. It is just that after the upgrade the corresponding output for each input will have changed.

 ### Before/After

<img width="985" alt="Screenshot 2024-01-02 at 15 12 14" src="https://github.com/oftherivier/fictional/assets/1731223/df64be59-705c-499c-8218-516e20e13855">

 ### Before on its own

<img width="975" alt="Screenshot 2024-01-02 at 15 13 20" src="https://github.com/oftherivier/fictional/assets/1731223/42de34c1-229d-444a-8593-b461a918619d">

 ### After on its own

<img width="1003" alt="Screenshot 2024-01-02 at 15 14 04" src="https://github.com/oftherivier/fictional/assets/1731223/b32952ea-09f2-4988-abcd-1b366b05fc1c">

### Script and values
https://gist.github.com/justinvdm/9494ed9ba728de802778832b3c9d640d