feat!: Replace phoneNumber {min,max} options with {min,max}Length
BREAKING CHANGE
* In the cases where the `prefixes` option is provided, copycat will now return a different output for the same input
* The `min` and `max` options have been removed in favour of `minLength` and `maxLength`

```ts
// before
copycat.phoneNumber(input, { min: 100000, max: 999999999 })

// after
copycat.phoneNumber(input, { minLength: 6, maxLength: 9 })
```

 ## TLDR
We discovered bugs in our implementation of phoneNumber, and in the process decided that `{min,max}Length` instead of `{min,max}` options would avoid the bugs altogether, while at the same time provide a better UX (more direct way of describing the intent, less room for human error)

```
copycat.phoneNumber(input, { minLength: 2, maxLength: 10 })
```

 ## Detailed context
 ### Part 1: The original issue
Currently, we support `min` and `max` options for `phoneNumber`, to be used like this:

```ts
copycat.phoneNumber(input, { min: 1000000, max: 9999999 })
```

To support this, for the case where users provided `prefixes` optios also, we had this code:

```
    const adjustedMin = Math.max(min, 10 ** (10 - prefixLength))
    const adjustedMax = Math.min(max, 10 ** (15 - prefixLength) - 1)
```

Consider the case where the given `min` is `100000000` and the given `max` is `999999999` - both of digit length 9. Here, there are two issues:
1. the hardcoded 10 and 15 do not match the lengths of the min and max given, which was 9
2. we're using Math.max() for the adjustedMin

Because of (1) and (2), we end up with an adjustedMin of 1000000000 (the computed value is larger than the users, so it gets taken), and a max of 999999999 (the user's max was used since it was smaller).

Here the max is actually smaller than the min, so we end up clamping the generated values to 1000000000 (i.e, the generated value will always be 1000000000 since it is now both the upper and lower bound).

 ### Part 1: Where the API change comes in

Working on solving the bugs outlined about lead me to try a test case that looks like this:

```ts
copycat.phoneNumber(input, { min: 21, max: 106 })
```

Then I realised that's a weird way to use phone number. What `min` and `max` are really meant to be given as, are values like `100000000` and `999999999`. In other words, they're really just meant for describing the desired min/max _lengths_ of the phone numbers. But then, why not just have the user be able to describe this directly?

```
copycat.phoneNumber(input, { minLength: 2, maxLength: 10 })
```

* The user can now express their intent directly
* We're also now giving less opportunity for the API to be used in a way that was not intended (just as I did with `{ min: 21, max: 106 }`)