chore: Change release approach for preview releases
Instead of needing to maintain `next` and `main` branches, we now have a single `main` branch, and a new `yarn release` script to wrap
up the release workflow. The script usage looks like this:

```
yarn release <major|minor|patch>
```

If the `major` is given as the version bump type, the script will also version release a preview version. For example, if the current version was 2.3.1, `yarn release major` would release two versions:
* 3.0.0-preview.0 - the same as 2.3.1, except @snaplet/copycat/next would allow the user to preview 3.0.0
* 3.0.0 - the actual major release
```

 ## Why the change?
The nature of copycat (since it is meant to map to the same output for the same input) means that changes to the code frequently cause breaking changes (since the output often changes when the code changes). This happens often enough that we need a more streamlined way of releasing these breaking changes.

The current approach is to maintain `next` and `main` branches and schedule breaking changes for a given date. This is problematic for a few reasons:
* Needing to maintain two branches requires unnecessary branch juggling whenever one of the branches changes
* Ideally users should be able to use major releases as soon as they would like to - without needing to change the way they use copycat and, without first waiting for a given date